pyproject.toml
[project]
name = "docrun"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "docling>=2.0.0",
    "google-genai>=0.8.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.12",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyserial>=3.5",
]

# Use CPU-only torch — avoids all CUDA packages (~3 GB saved)
[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
